Feat: GNB 확장 및 경로에 따른 스타일 분리
경로에 따라 GNB에 다른 스타일을 적용하도록 리팩토링.
- mypage 관련 경로와 그 외 기본 경로의 gnb 높이, 패딩 다르게 설정.
- mypage 관련 경로에서는 알림 아이콘과 프로필 아이콘을 표시하고, 그 외 기본 경로에서는 로그인 버튼을 노출.
- 알림 아이콘과 프로필 아이콘을 img tag에서 SVG 컴포넌트로 수정
- 알림 아이콘과 프로필 아이콘에 onClick 이벤트를 추가해 페이지 이동 기능 구현.
