--- .devcontainer/frontend/src/components/DiaryHeader.tsx
@@ -1,54 +1,0 @@
-import { useNavigate } from "react-router-dom";
-import img from "../assets/image 13.svg";
-
-const DiaryHeader = () => {
-  const navigate = useNavigate();
-
-  const handleAlarmClick = () => {
-    navigate("/alert");
-  };
-
-  const handleProfileClick = () => {
-    navigate("/mypage");
-  };
-
-  return (
-    <header className="flex justify-between items-center w-full h-[62px] px-4 md:px-20 py-4 bg-white border-b border-[#e2e2e2]">
-      <p className="text-2xl font-medium text-[#1d1b20]">Logo</p>
-      <div className="flex items-center space-x-4">
-        <svg
-          width={24}
-          height={24}
-          viewBox="0 0 24 24"
-          fill="none"
-          xmlns="http://www.w3.org/2000/svg"
-          className="w-6 h-6 cursor-pointer"
-          onClick={handleAlarmClick}
-        >
-          <g clipPath="url(#clip0_1793_127)">
-            <path
-              fillRule="evenodd"
-              clipRule="evenodd"
-              d="M6.70221 2.62368C8.10763 1.21827 10.0138 0.428711 12.0014 0.428711C13.9889 0.428711 15.8951 1.21827 17.3004 2.62368C18.706 4.02911 19.4954 5.93527 19.4954 7.92285C19.4954 8.85422 19.6475 9.74431 19.8902 10.6889C19.9523 10.8807 20.0198 11.061 20.0918 11.2304C20.189 11.4589 20.3855 11.6562 20.7629 11.8977C20.9104 11.992 21.0612 12.0787 21.2327 12.1773L21.3929 12.2696C21.6211 12.4017 21.8779 12.5551 22.1232 12.7366C23.0099 13.3931 23.3359 14.4047 23.2919 15.3168C23.249 16.2043 22.8511 17.1294 22.1087 17.6876C22.0985 17.6957 22.0871 17.7044 22.0744 17.7139C22.0442 17.7365 22.0065 17.7629 21.9602 17.7925C21.8678 17.852 21.7413 17.9245 21.5726 18.0053C21.2357 18.1666 20.7278 18.3619 19.9807 18.5508C18.4895 18.9276 16.0195 19.2859 12.0014 19.2859C7.98321 19.2859 5.51318 18.9276 4.02201 18.5508C3.27494 18.3619 2.76703 18.1666 2.43009 18.0053C2.26139 17.9245 2.13494 17.852 2.04251 17.7925C1.99627 17.7629 1.95856 17.7365 1.92832 17.7139C1.91562 17.7044 1.90422 17.6957 1.89409 17.6876C1.15161 17.1294 0.753781 16.2043 0.710924 15.3168C0.666882 14.4047 0.992811 13.3931 1.87957 12.7366L2.38957 13.4256L1.87957 12.7366C2.12481 12.5551 2.38156 12.4017 2.60971 12.2696L2.77005 12.1772L2.77009 12.1772C2.94153 12.0787 3.09242 11.992 3.23983 11.8977C3.61723 11.6562 3.81366 11.4589 3.91081 11.2304C4.25761 10.4149 4.50722 9.33755 4.50722 7.92285C4.50722 5.93527 5.29679 4.02911 6.70221 2.62368ZM10.0099 20.5452C9.6765 20.4317 9.30772 20.5337 9.08001 20.8023C8.85232 21.0709 8.81203 21.4515 8.97846 21.7618C9.55523 22.8372 10.6919 23.5716 12.0014 23.5716C13.3109 23.5716 14.4476 22.8372 15.0243 21.7618C15.1908 21.4515 15.1505 21.0709 14.9228 20.8023C14.6951 20.5337 14.3263 20.4317 13.993 20.5452C13.5588 20.6928 12.9174 20.8057 12.0014 20.8057C11.0855 20.8057 10.444 20.6928 10.0099 20.5452Z"
-              fill="#E2E2E2"
-            />
-            <circle cx={20} cy={4} r="3.5" fill="#2768FF" stroke="white" />
-          </g>
-          <defs>
-            <clipPath id="clip0_1793_127">
-              <rect width={24} height={24} fill="white" />
-            </clipPath>
-          </defs>
-        </svg>
-        <img
-          src={img}
-          alt="Profile"
-          className="w-10 h-10 rounded-full object-cover border border-[#e2e2e2] cursor-pointer"
-          onClick={handleProfileClick}
-        />
-      </div>
-    </header>
-  );
-};
-
-export default DiaryHeader;
